check if outputs exitt

+++ src/scripts/seqbox_filehandling.py
@@ -97,37 +97,38 @@ def run_add_artic_consensus_to_filestructure(args):
         group_name = sample.sample_source.projects[0].groups.group_name
         target_dir = os.path.join(config['seqbox_directory'], group_name, f"{rs.readset_identifier}-{sample.sample_identifier}", "artic_pipeline")
         target_consensus = os.path.join(target_dir, f"{rs.readset_identifier}-{sample.sample_identifier}.artic.consensus.fasta")
-        if args.manual_artic_pipeline is True:
-            source_consensus = glob.glob(
-                f"{args.consensus_genomes_parent_dir}/{args.readset_batch_name}_{rs.readset_nanopore.barcode}.consensus.fasta")
-        else:
-            source_consensus = glob.glob(f"{args.consensus_genomes_parent_dir}/{args.readset_batch_name}_{rs.readset_nanopore.barcode}/*consensus.fasta")
-        if len(source_consensus) == 1:
-            if not os.path.isdir(target_dir):
-                os.mkdir(target_dir)
-            os.symlink(source_consensus[0], target_consensus)
-            print(f"Linked {source_consensus[0]} to {target_consensus}")
-        elif len(source_consensus) == 0:
-            print(f"No consensus genome at {args.consensus_genomes_parent_dir}/{args.readset_batch_name}_{rs.readset_nanopore.barcode}*")
-        elif len(source_consensus) > 1:
-            print(f"More than one consensus genome at {args.consensus_genomes_parent_dir}/*{rs.readset_nanopore.barcode}*")
-
-        if args.manual_artic_pipeline is True:
-            source_bam = glob.glob(
-                f"{args.consensus_genomes_parent_dir}/{args.readset_batch_name}_{rs.readset_nanopore.barcode}.primertrimmed.rg.sorted.bam")
-        else:
-            source_bam = glob.glob(f"{args.consensus_genomes_parent_dir}/{args.readset_batch_name}_{rs.readset_nanopore.barcode}/*primertrimmed.rg.sorted.bam")
+        if not os.path.isfile(target_consensus):
+            if args.manual_artic_pipeline is True:
+                source_consensus = glob.glob(
+                    f"{args.consensus_genomes_parent_dir}/{args.readset_batch_name}_{rs.readset_nanopore.barcode}.consensus.fasta")
+            else:
+                source_consensus = glob.glob(f"{args.consensus_genomes_parent_dir}/{args.readset_batch_name}_{rs.readset_nanopore.barcode}/*consensus.fasta")
+            if len(source_consensus) == 1:
+                if not os.path.isdir(target_dir):
+                    os.mkdir(target_dir)
+                os.symlink(source_consensus[0], target_consensus)
+                print(f"Linked {source_consensus[0]} to {target_consensus}")
+            elif len(source_consensus) == 0:
+                print(f"No consensus genome at {args.consensus_genomes_parent_dir}/{args.readset_batch_name}_{rs.readset_nanopore.barcode}*")
+            elif len(source_consensus) > 1:
+                print(f"More than one consensus genome at {args.consensus_genomes_parent_dir}/*{rs.readset_nanopore.barcode}*")
 
         target_bam = os.path.join(target_dir, f"{rs.readset_identifier}-{sample.sample_identifier}.artic.bam")
-        if len(source_bam) == 1:
-            if not os.path.isdir(target_dir):
-                os.mkdir(target_dir)
-            os.symlink(source_bam[0], target_bam)
-            print(f"Linked {source_bam[0]} to {target_bam}")
-        elif len(source_bam) == 0:
-            print(f"No bam at {args.consensus_genomes_parent_dir}/{args.readset_batch_name}_{rs.readset_nanopore.barcode}*")
-        elif len(source_bam) > 1:
-            print(f"More than one bam at {args.consensus_genomes_parent_dir}/{args.readset_batch_name}_{rs.readset_nanopore.barcode}*")
+        if not os.path.isfile(target_bam):
+            if args.manual_artic_pipeline is True:
+                source_bam = glob.glob(
+                    f"{args.consensus_genomes_parent_dir}/{args.readset_batch_name}_{rs.readset_nanopore.barcode}.primertrimmed.rg.sorted.bam")
+            else:
+                source_bam = glob.glob(f"{args.consensus_genomes_parent_dir}/{args.readset_batch_name}_{rs.readset_nanopore.barcode}/*primertrimmed.rg.sorted.bam")
+            if len(source_bam) == 1:
+                if not os.path.isdir(target_dir):
+                    os.mkdir(target_dir)
+                os.symlink(source_bam[0], target_bam)
+                print(f"Linked {source_bam[0]} to {target_bam}")
+            elif len(source_bam) == 0:
+                print(f"No bam at {args.consensus_genomes_parent_dir}/{args.readset_batch_name}_{rs.readset_nanopore.barcode}*")
+            elif len(source_bam) > 1:
+                print(f"More than one bam at {args.consensus_genomes_parent_dir}/{args.readset_batch_name}_{rs.readset_nanopore.barcode}*")
 
 
 def run_command(args):
